dynamic handlesubmit funtion for all input components
