+++ src/App.js
@@ -5,13 +5,11 @@ import LogIn from './components/log-in';
 import Register from './components/register';
 import Dashboard from './pages/dashboard';
 import CreateRecipePage from './pages/create-recipe-page';
-import CreateRecipe from './components/create-recipe';
 
 function App() {
   return (
     <div className="App">
-      <CreateRecipe />
-      {/* <CreateRecipePage /> */}
+      <CreateRecipePage />
       {/* <Switch>
         <Route exact path="/" component={LogIn} />
         <PrivateRoute exact path="/dashboard" component={Dashboard}/>
