+++ src/components/recipe-input.js
@@ -5,10 +5,7 @@ import { makeStyles } from '@material-ui/core/styles';
 import TextField from '@material-ui/core/TextField';
 
 const useStyles = makeStyles((theme) => ({
-  input: {
-    width: '40%',
-    height: '15%',
-  }
+
 }));
 
 const RecipeInput = (props) => {
